Update build.sh
more testing

+++ build.sh
@@ -27,10 +27,12 @@ CYAN=$(tput setaf 6)
 RESET=$(tput sgr0)
 
 #Update check
-git stash push
-git stash drop
-git pull https://github.com/gunvalk/sm64pcBuilder
-./build.sh
+if [[ `git status --porcelain` ]]; then
+	git stash push
+	git stash drop
+	git pull https://github.com/gunvalk/sm64pcBuilder
+	./build.sh
+fi
 
 #Update message
 cat<<EOF
@@ -39,15 +41,7 @@ cat<<EOF
     ${YELLOW}------------------------------${RESET}
     ${CYAN}Updates:
 
-    You will no longer need to update your build.sh file manually.                    
-    There will now be a sm64pcBuilder folder on your C drive. 
-    This is the folder where your build.sh files will generate,
-    as well as your sm64pc-master or sm64pc-nightly folders.
-    You can delete any build.sh file that is outside of
-    the sm64pcBuilder. Your old sm64pc-master or sm64pc-nightly 
-    folders are in the same location as they were. Either move
-    them to the sm64pcBuilder folder, or select yes to updating
-    when the green message appears to generate new folders.                    
+    I dont feel so good mr stark                                    
 
     ${RESET}${YELLOW}------------------------------${RESET}
     ${CYAN}build.sh Update 15${RESET}
